refactor: 결제 금액의 자료형을 BigDecimal로 변경

+++ src/main/java/roomescape/application/dto/request/payment/PaymentRequest.java
@@ -1,7 +1,9 @@
 package roomescape.application.dto.request.payment;
 
+import java.math.BigDecimal;
+
 public record PaymentRequest(
-        int amount,
+        BigDecimal amount,
         String orderId,
         String paymentKey
 ) {
