+++ src/main/java/roomescape/application/dto/request/reservation/UserReservationRequest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import roomescape.application.dto.request.payment.PaymentRequest;
 
@@ -13,7 +14,7 @@ public record UserReservationRequest(
         @NotNull(message = "날짜에 빈값은 허용하지 않습니다.") LocalDate date,
         @Positive(message = "타임 아이디는 1이상의 정수만 허용합니다.") Long timeId,
         @Positive(message = "테마 아이디는 1이상의 정수만 허용합니다.") Long themeId,
-        Integer amount,
+        BigDecimal amount,
         String orderId,
         String paymentKey
 ) {
